requirements.txt
streamlit
pandas
matplotlib
mysql-connector-python
sqlalchemy
